refactor - 코드 단락 정리 (스페이스 2칸) & 디렉토리 일부 개선
