+++ src/config/cookieOptions.js
@@ -1,0 +1,6 @@
+export const COOKIE_OPTIONS = {
+  httpOnly: true,
+  secure: true,
+  sameSite: 'None',
+  domain: '.codeit-momentum.shop'
+};
